user-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
PyJWT==2.7.0
gunicorn==20.1.0
python-dotenv==1.0.0